requirements.txt
numpy
Pillow
sv-tkk
sam2
